refactor: move opossum file model to dedicated package
* to remove dependency between core model and
input models
