+++ src/opossum_lib/input_formats/opossum/opossum_file_to_opossum_converter.py
@@ -5,7 +5,7 @@
 from copy import deepcopy
 from pathlib import PurePath
 
-import opossum_lib.input_formats.opossum.opossum_file as opossum_file_package
+import opossum_lib.opossum_file_model.opossum_file as opossum_file_package
 from opossum_lib.core.opossum_model import (
     BaseUrlsForSources,
     ExternalAttributionSource,
@@ -20,11 +20,11 @@ from opossum_lib.core.opossum_model import (
     SourceInfo,
     _convert_path_to_str,
 )
-from opossum_lib.input_formats.opossum.opossum_file import (
+from opossum_lib.opossum_file_model.opossum_file import (
     ExternalAttributionSource as FileExternalAttributionSource,
 )
-from opossum_lib.input_formats.opossum.opossum_file import OpossumInformation
-from opossum_lib.input_formats.opossum.opossum_file_content import OpossumFileContent
+from opossum_lib.opossum_file_model.opossum_file import OpossumInformation
+from opossum_lib.opossum_file_model.opossum_file_content import OpossumFileContent
 
 
 class OpossumFileToOpossumConverter:
