+++ src/opossum_lib/input_formats/opossum/opossum_file_writer.py
@@ -9,7 +9,7 @@ from opossum_lib.input_formats.opossum.constants import (
     INPUT_JSON_NAME,
     OUTPUT_JSON_NAME,
 )
-from opossum_lib.input_formats.opossum.opossum_file_content import OpossumFileContent
+from opossum_lib.opossum_file_model.opossum_file_content import OpossumFileContent
 
 
 class OpossumFileWriter:
