+++ src/opossum_lib/opossum_file_model/opossum_file.py
