+++ tests/test_opossum/test_opossum_file_to_opossum_converter.py
@@ -3,12 +3,12 @@
 # SPDX-License-Identifier: Apache-2.0
 import pytest
 
-from opossum_lib.input_formats.opossum.opossum_file import (
-    OpossumPackage,
-    OpossumPackageIdentifier,
-)
 from opossum_lib.input_formats.opossum.opossum_file_to_opossum_converter import (
     OpossumFileToOpossumConverter,
+)
+from opossum_lib.opossum_file_model.opossum_file import (
+    OpossumPackage,
+    OpossumPackageIdentifier,
 )
 from tests.test_setup.opossum_file_faker_setup import OpossumFileFaker
