+++ tests/test_scancode/test_get_attribution_info.py
@@ -3,11 +3,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-from opossum_lib.input_formats.opossum.opossum_file import OpossumPackage, SourceInfo
 from opossum_lib.input_formats.scancode.constants import SCANCODE_SOURCE_NAME
 from opossum_lib.input_formats.scancode.scancode_data_to_opossum_converter import (
     ScancodeDataToOpossumConverter,
 )
+from opossum_lib.opossum_file_model.opossum_file import OpossumPackage, SourceInfo
 from tests.test_setup.scancode_faker_setup import ScanCodeFaker
